move fetcher to a separate file
